refactor(staff): update avatar source, add staff type
