+++ components/ui/card3d/CardBody.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { cn } from '~~/lib/utils'
+import { cn } from '~/lib/utils'
 
 defineProps({
   class: String,
