+++ constants/staff.ts
@@ -1,0 +1,126 @@
+import type { Staff } from '~/types/staff'
+
+export const staffData: Staff[] = [
+  {
+    name: '高見龍',
+    jobTitle: '掃地兼敲鐘',
+    avatar: '/staff/avatar_eddie.jpg',
+    socialLinks: [],
+  },
+  {
+    name: '廖洧杰',
+    jobTitle: '專業推坑王',
+    avatar: '/speakers/avatar_placeholder.svg',
+    socialLinks: [
+      {
+        type: 'facebook',
+        url: 'https://www.facebook.com/sfismy',
+      },
+    ],
+  },
+  {
+    name: '理查哥',
+    jobTitle: 'UX 土地公',
+    avatar: '/staff/avatar_richard.jpg',
+    socialLinks: [],
+  },
+  {
+    name: '陳彥宇',
+    jobTitle: '年度視覺擔當',
+    avatar: '/speakers/avatar_placeholder.svg',
+    socialLinks: [],
+  },
+  {
+    name: '豪萱',
+    jobTitle: '設計小夥伴',
+    avatar: '/staff/avatar_birteliu.jpg',
+    socialLinks: [
+      {
+        type: 'facebook',
+        url: 'https://www.facebook.com/zyi870213',
+      },
+    ],
+  },
+  {
+    name: 'EG',
+    jobTitle: '設計小夥伴',
+    avatar: '/speakers/avatar_placeholder.svg',
+    socialLinks: [
+      {
+        type: 'facebook',
+        url: 'https://www.facebook.com/eg.pan',
+      },
+    ],
+  },
+  {
+    name: 'Nina',
+    jobTitle: '設計小夥伴',
+    avatar: '/staff/avatar_nina.jpg',
+    socialLinks: [],
+  },
+  {
+    name: '楊正弘',
+    jobTitle: '設計小夥伴',
+    avatar: '/speakers/avatar_placeholder.svg',
+    socialLinks: [
+      {
+        type: 'instagram',
+        url: 'https://www.instagram.com/helloscott23050000/',
+      },
+    ],
+  },
+
+  {
+    name: '哲樺｜Frank',
+    jobTitle: '開發小農夫',
+    avatar: '/staff/avatar_frank.jpg',
+    socialLinks: [
+      {
+        type: 'linkedin',
+        url: 'https://www.linkedin.com/in/%E5%93%B2%E6%A8%BA-frank/',
+      },
+      {
+        type: 'instagram',
+        url: 'https://www.instagram.com/chu_chehua/',
+      },
+      {
+        type: 'website',
+        url: 'https://frank-code-world.com.tw/',
+      },
+    ],
+  },
+  {
+    name: 'Tim',
+    jobTitle: '開發小農夫',
+    avatar: '/staff/avatar_tim.jpg',
+    socialLinks: [
+      {
+        type: 'linkedin',
+        url: 'https://www.linkedin.com/in/timinitime/',
+      },
+      {
+        type: 'website',
+        url: 'https://timemo.vercel.app/',
+      },
+    ],
+  },
+  {
+    name: '薛羽婷',
+    jobTitle: '設計軌講者獵人',
+    avatar: '/staff/avatar_sha.jpg',
+    socialLinks: [],
+  },
+  {
+    name: 'Sabrina',
+    jobTitle: '活動工頭',
+    avatar: '/speakers/avatar_placeholder.svg',
+    socialLinks: [],
+  },
+
+  {
+    name: 'Melissa',
+    jobTitle: '錢錢擔當',
+    avatar: '/speakers/avatar_placeholder.svg',
+    socialLinks: [],
+  },
+]
