+++ types/common.ts
@@ -1,0 +1,1 @@
+export type SocialLinkType = 'facebook' | 'x' | 'instagram' | 'website' | 'youtube' | 'thread' | 'medium' | 'linkedin'
