+++ types/staff.ts
@@ -1,0 +1,11 @@
+import type { SocialLinkType } from './common'
+
+export interface Staff {
+  name: string
+  jobTitle?: string
+  avatar: string
+  socialLinks: {
+    type: SocialLinkType
+    url: string
+  }[]
+}
